Refactor: Correct Go import paths in backend modules
I removed the erroneous '/internal' path segment from import paths
for modules located directly under the 'backend' directory.

This change aligns the import paths with the actual file structure
and the Go module path defined in `go.mod` (`github.com/water-classroom/backend`).

For example, an import like:
`github.com/water-classroom/backend/internal/app`
has been changed to:
`github.com/water-classroom/backend/app`

This correction was applied to all relevant Go files within the
`backend` directory, including application code, tests, and
commented-out code.
