+++ backend/assessment/handlers.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/water-classroom/backend/internal/app"
+	"github.com/water-classroom/backend/app"
 	"go.uber.org/zap"
 )
