+++ backend/auth/jwt.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/golang-jwt/jwt/v5"
-	"github.com/water-classroom/backend/internal/app" // Adjusted
+	"github.com/water-classroom/backend/app" // Adjusted
 	"go.uber.org/zap"                                                  // Added
 )
