+++ backend/payment/routes.go
@@ -2,7 +2,7 @@ package payment
 
 import (
 	"github.com/gin-gonic/gin"
-	// "github.com/water-classroom/water-classroom-monolith/internal/app"
+	"github.com/water-classroom/backend/app"
 	"go.uber.org/zap"
 )
